requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.0
cycler==0.12.1
fonttools==4.60.2
hdbscan==0.8.41
idna==3.11
importlib_resources==6.5.2
joblib==1.5.3
kiwisolver==1.4.7
matplotlib==3.9.4
numpy<2.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
quadprog==0.1.13
requests==2.32.5
scikit-learn==1.6.1
scikit-optimize>=0.10.2
bottleneck>=1.3.6
scipy==1.13.1
six==1.17.0
sklearn==0.0
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.3
urllib3==2.6.2
wget==3.2
zipp==3.23.0
